// src/tasks/enum/task.enum.ts
export enum TaskLabel {
  clientregistration_contact_details = 'clientregistration_contact_details',
  clientregistration_financial_info = 'clientregistration_financial_info',
  clientregistration_expereince_assessment = 'clientregistration_expereince_assessment',
  clientregistration_declaration_signature = 'clientregistration_declaration_signature',
  clientregistration_identity_document_upload = 'clientregistration_identity_document_upload',
  clientregistration_address_document_upload = 'clientregistration_address_document_upload',
  clientregistration_payment_document_upload = 'clientregistration_payment_document_upload',
  clientregistration_update_kyc_docs = 'clientregistration_update_kyc_docs',
}