// src/admin/active-log/active-log.type.ts
export enum performerType {
  USER = 'User',
  OPERATOR = 'Operator',
  SYSTEM = 'System',
  CRMAPI = 'CrmAPI',
}

export enum actionType {
  DETAILS_UPDATED = 'DetailsUpdated',
  RECORD_CREATED = 'RecordCreated',
  RECORD_DELETED = 'RecordDeleted',
  NOTE_ADDED = 'NoteAdded',
  PASSWORD_CHANGED = 'PasswordChanged',
  DEVICE_ADDED = 'DeviceAdded',
  STUB = 'Stub',
  RECORD_DOWNLOADED = 'RecordDownloaded',
}

export enum entityType {
  USER = 'User',
  BROKER_USER = 'BrokerUser',
  OPERATOR = 'Operator',
  KYC_DOCUMENT = 'KycDocument',
  TRANSACTION = 'Transaction',
  AFFILIATE = 'Affiliate',
  CALLBACK = 'Callback',
  BROKER_BANKING = 'BrokerBanking',
  BROKER = 'Broker',
  BROKER_COUNTRY = 'BrokerCountry',
  OFFICE = 'Office',
  COMMUNICATION_INTEGRATION = 'CommunicationIntegration',
  TEMPLATE = 'Template',
  EVENT_COMMUNICATION_TYPE_REL = 'EventCommunicationTypeRel',
  COMMUNICATION_PARAMETER = 'CommunicationParameter',
  TRANSACTION_LIMITS = 'TransactionLimits',
  ROLE = 'Role',
  AFFILIATE_BANKING = 'AffiliateBanking',
  APP = 'App',
  DESK_STATUS_REROUTE_RULE = 'DeskStatusRerouteRule',
  NINJA_RULE = 'NinjaRule',
  MENUITEM = 'MenuItem',
  DESK_RULE = 'DeskRule',
  CRMAPI_TOKEN = 'CrmApiToken',
  AFFILIATE_PAYOUT = 'AffiliatePayout',
  WEB_ASSET = 'WebAsset',
  NINJA_PARAMETER = 'NinjaParameter',
  PARAMETER = 'Parameter',
  PSP = 'Psp',
  USER_BANKING = 'UserBanking',
  TRANSLATION_KEY_LANG = 'TranslationKeyLang',
  TRANSLATION_KEY = 'TranslationKey',
  USER_SITE_LINKS_VIEW = 'UserSiteLinksView',
  PSP_RULE = 'PspRule',
  CONTENT_LINK = 'ContentLink',
  USER_SITE_PSPS_VIEW = 'UserSitePspsView',
  CREDIT = 'Credit',
  ADVERTISER_PAYMENT = 'AdvertiserPayment',
  ADVERTISER = 'Advertiser',
  LEAD = 'Lead',
  CALL_LOGS = 'CallLogs',
  TASK = 'Task',
  MEETING = 'Meeting',
  TICKET = 'Ticket',
  MT5_ACCOUNT = 'MT5_Account',
  TICKET_COLLABORATOR = 'TicketCollaborator',
  TICKET_NOTES = 'TicketNotes',
  OPPORTUNITY = 'Opportunity',
  CLIENT_TASK = 'ClientTask',
  LABEL = 'Label',
  REGULATION = 'Regulation',
  NOTE = 'Note',
  REGULATION_RULE = 'RegulationRule',
  REGULATION_EVENT = 'RegulationEvent',
  REGULATION_MAPPING = 'RegulationMapping',
  LIST_NAME = 'ListName',
  LIST_COLUMNS_FILTER = 'ListColumnFilter',
  LIST_VIEWS_FILTER = 'ListViewsFilter',
  LIST_VIEW_COLUMN = 'ListViewColumn',
  EMAIL_EVENT = 'EmailEvent',
  BODY_CONTENT = 'BodyContent',
  LAYOUT = 'Layout',
  DESK = 'Desk',
  AGGREGATOR='Aggregator',
  COMPANY_BANKS='CompanyBank',
  MERGE_TICKET = 'MergeTicket',
  IB_COMMISSION_PROFILE = 'IBCommissionProfile',
  IB_COMMISSION_PROFILE_CONFIG = 'IBCommissionProfileConfig',
  EVENT_REGISTRATION = 'EventRegistration',
  PSP_COUNTRY_PRIORITY_CONFIG='PspCountryPriorityConfig',
  REFERRAL ='Referral',
  REFERRAL_PROGRAM = 'ReferralProgram',
  BONUS = 'Bonus',
}